Added new mechanisms to handle different backends
Added a bunch of defines on config.hpp that can be used to handle diffeerent backends. Also
created a define named GDX_BACKEND_PREFIX which will store a string referencing the name of the backend
like "ios","linux", "android" and so on, so it's easier to handle per backend configuration strings.