Many improvements regarding support to c++11 features
I've added support to lambdas when posting to the main thread, making this operation
a breeze when on multithreaded operations. Fixed the log output on Android devices
to print only the filename without full path.
Changed the Thread class behaviour, not referencing the thread pointer with a shared_ptr
will not abort it's execution. Now it will run until it stops, and if there's no reference
to it it will delete itself.