Updated the android toolchain to run up to r8d. Changes regarding generated files
All files that need to be configured (config.hpp, gl.hpp) are now bound to the build directory
instead of the source directory. This turns the cross-platform management much easier, but requires
the user to run

initialize_gdx()

after finding the package. It also will set the GDX_BACKEND_<PLATFORM> variables so it can be
useful to the lib user to know in which plaftorm he is.